common js functions moved into common.js file; scripts.js renamed to index.js; scripts loaded in *.html files with type='module'
