+++ client/comments.js
@@ -1,3 +1,5 @@
+import { makeRequestToUnsplash, requestAccessKey, createCardVotesAndButtons } from "./common.js";
+
 const divContent = document.getElementById("content");
 // const divLeaveComment = document.getElementById("leave-a-comment");
 // const divComments = document.getElementById("comments");
@@ -5,36 +7,15 @@ const divContent = document.getElementById("content");
 createPageContent();
 
 async function createPageContent() {
-  const apiResponse = await makeRequestToUnsplash();
-  divContent.prepend(createPictureCard(apiResponse));
+  const ACCESS_KEY = await requestAccessKey();
+  const currentUrl = window.location.href;
+  const photoId = String(currentUrl).split("=")[1];
+  const url = `https://api.unsplash.com/photos/${photoId}?client_id=${ACCESS_KEY}`;
+
+  const jsonData = await makeRequestToUnsplash(url);
+  divContent.prepend(createPictureCard(jsonData));
   // divContent.appendChild(createLeaveCommentCard(apiResponse));
   // divContent.appendChild(createCommentsCards(apiResponse));
-}
-
-async function makeRequestToUnsplash() {
-  const ACCESS_KEY = await requestAccessKey();
-  const url = window.location.href;
-  const photoId = String(url).split("=")[1];
-  const requestUrl = `https://api.unsplash.com/photos/${photoId}?client_id=${ACCESS_KEY}`;
-  try {
-    const response = await fetch(requestUrl);
-    const jsonData = await response.json();
-    return await jsonData;
-  } catch (error) {
-    console.error("The request to Unsplash is rejected!", error);
-    return error;
-  }
-}
-
-async function requestAccessKey() {
-  try {
-    const response = await fetch("http://localhost:8080");
-    const jsonData = await response.json();
-    return await jsonData.access_key;
-  } catch (error) {
-    console.error("The attempt to get API key is rejected!", error);
-    return error;
-  }
 }
 
 function createPictureCard(imageObj) {
@@ -61,35 +42,13 @@ function createPictureCard(imageObj) {
   const image = document.createElement("img");
   image.src = imageObj.urls.regular;
   image.alt = imageObj.alt_description;
-  image.classList.add("card-img-top", "img-fluid", "px-3");
+  image.classList.add("card-img-top", "img-fluid", "ps-3");
   imageDiv.appendChild(image);
   // buttons
-  const buttonsDiv = document.createElement("div");
-  buttonsDiv.classList.add("m-auto", "me-3", "buttons-votes", "text-center");
-  // plus button
-  const plusButton = document.createElement("button");
-  plusButton.type = "button";
-  plusButton.classList.add("btn", "btn-sm", "btn-outline-secondary", "vote-btn");
-  const plusSign = document.createTextNode("+");
-  plusButton.appendChild(plusSign);
-  // likes
-  const votesDiv = document.createElement("div");
-  votesDiv.classList.add("votes", "d-block", "pt-2", "pb-2", "text-center");
-  const votesNumber = document.createTextNode(imageObj.likes);
-  votesDiv.appendChild(votesNumber);
-  // minus button
-  const minusButton = document.createElement("button");
-  minusButton.type = "button";
-  minusButton.classList.add("btn", "btn-sm", "btn-outline-secondary", "vote-btn");
-  const minusSign = document.createTextNode("-");
-  minusButton.appendChild(minusSign);
-  // construct buttons div
-  buttonsDiv.appendChild(plusButton);
-  buttonsDiv.appendChild(votesDiv);
-  buttonsDiv.appendChild(minusButton);
+  const buttons = createCardVotesAndButtons(imageObj)
   // construct picture and buttons div
   divPictureAndButtons.appendChild(imageDiv);
-  divPictureAndButtons.appendChild(buttonsDiv);
+  divPictureAndButtons.appendChild(buttons); 
   // card body
   const cardBodyDiv = document.createElement("div");
   cardBodyDiv.classList.add("card-body", "pt-1", "pb-2");
