+++ client/index.html
@@ -182,7 +182,7 @@
 
         <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha1/dist/js/bootstrap.bundle.min.js" 
         integrity="sha384-w76AqPfDkMBDXo30jS1Sgez6pr3x5MlQ1ZAGC+nuZB+EYdgRZgiwxhTBTkF7CXvN" crossorigin="anonymous"></script>
-        <script src="scripts.js"></script>
+        <script src="index.js", type="module"></script>
         <!-- <script>  includeHTML(); </script>  -->
     
     </body>
