+++ client/package.json
@@ -1,6 +1,7 @@
 {
   "name": "client",
   "version": "1.0.0",
+  "type": "module",
   "description": "client for gallery app",
   "main": "comments.js",
   "scripts": {
